+++ sktime/annotation/hmm_learn/__init__.py
@@ -11,4 +11,4 @@ __all__ = [
 
 from sktime.annotation.hmm_learn.base import BaseHMMLearn
 from sktime.annotation.hmm_learn.gaussian import GaussianHMM
-from sktime.annotation.hmm_learn.gmmhmm import GMMHMM
+from sktime.annotation.hmm_learn.gmm import GMMHMM
